Add -sample to config files; ignore config files
This commit includes two changes to the WPSK setup:

1. Both `local-config.php` and `wp-config.php` are ignored in the
repository by default. `local-config.php` should always be ignored as
described by Mark Jaquith. `wp-config.php` is now ignored by default to
reduce the likelihood of accidentally pushing it to a cloud like GitHub.
Pushing `wp-config.php` to an remote now requires manually removing it
from `.gitignore`.

2. `wp-config.php` and `local-config.php` now come with `-sample`
attached to the filenames. This allows the files to be stored in the
WPSK in the first place, now that the non-`sample` files are ignored via
`.gitignore`. It also reinforces the need to edit them and to think
about whether to ignore `wp-config.php` in your repository.